Many updates
Animations were not showing up in the previous commit so I grabbed a version that does work and is live now

https://raffle-spinner.000webhostapp.com/

And tried to find differences that would be causing issues. I was not successful so I copied the working version into this repo.

If there are any issues, please let me know