Adjusted Fonts and Edge Case
Commented out in css file for the body to have all Upper Case text.

Added in Roboto from google (bootstrap is overwriting I believe)

Fixed an edge case for which column to display: if the user wants to display all columns which means the player[1] value would be written as undefined in the Winner spreadsheet. This now just leaves a blank string.

Adjusted the modal that displays the sample of rows so that it will "show" after the [readxlsxfile] function has read the file.