Entropy Increased - Error Handling
Added the Crypto API to increase the randomness of the drawings. The names will still be randomized through Math.random() but that is only for show.

The winner is picked through the cryptoWinner() function and added to the end of the list.

Exclusions work and so does console logging.

Added more try-catch blocks and updated some wording to be more descriptive.